api/requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
google-cloud-bigquery>=3.10.0
cachetools>=5.3.3
